Allow matrix4x4 to interface with different arithmetic types
Previously, if you wanted to translate a Matrix4x4f by 10,10 you had to supply the 10,10 as 10.0f,10.0f, since the type of the parameters had to be the same as the type of the data in the data structure. Now it will allow any arithmetic type. So if you want to translate it by 10,10 you can just do that. It'll work the same as when multiplying/dividing/adding/subtracitng the data structure's float by the parameter's int. In other words, the int gets cast to a float implicitly by the arithmetic operators in the function's implementation.