Fix ambiguity of std::abs call
MinGW was giving an error on this code because it didn't know what version of std::abs to call with uint32_t.
I'm going out on a limb here to say that because we're right-shifting this integer by 24 bits, it's got to be able to fit inside a short. So short should be sufficient.